Update README.md
Stream 1 Project - Update 18.04
After an unintended break from being able to code and carry on the course, with fresh eyes I have abandoned the ‘band’ website, mainly as I don’t like the music as much as others and as I am designing it I thought why not change to music I like listening to, so I am changing, slightly. Premise of the site will remain the same, what it will do functionality wise, music, videos, bookings etc, however it will revolve around a DJ, Norman Cook aka Fat Boy Slim.
The criteria for the project was;
'Create A Website For A Band'
'The band is a 1960’s rock band and have around 50 years experience of performing live at numerous events around the world. You have been given the following requirements after interviews with the client’s representatives:
Their primary target audience are their fans and potential fans who wish to use the site to see and hear clips from their back catalogue, and any new material as it becomes available.
In addition, the band would like to use the site to showcase their music and publicize their availability to perform at events such as weddings and Christmas parties.
The band has provided you, the developer, with the following assets that they would like to showcase on their website:
Photos of the band members, A video clip, Audio clips
The band would also like to publicize upcoming gigs in the future. They would like users to be able to contact them with potential work related leads. The site should contain the ability to check on the band’s availability for future bookings and costs associated with various types of booking. To do this the band would like you to create a contact form which will take all of the details, such as, name, email, date, event type. Based on this data, the site will inform the user whether or not the date is available and the cost estimate for the event. All of this should be done using client-side JavaScript (no backend.)
In addition, they are in the process of creating a social media presence and would like to add links to their Facebook, Twitter and YouTube pages.'
The aim is to demonstrate use of HTML5 markup, CSS, JavaScript logic, jQuery/Angular
Release History
0.1 - Work In Progress
[I learned all this from Code Institute!](http://codeinstitute.net)
All work and views are my own